reademe

+++ BillingService/src/main/java/org/example/billingservice/Repository/ProductItemsRepository.java
@@ -3,5 +3,5 @@ package org.example.billingservice.Repository;
 import org.example.billingservice.Entitie.ProductItems;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductItemsRepository extends JpaRepository<ProductItems, Integer> {
+public interface ProductItemsRepository extends JpaRepository<ProductItems, Long> {
 }
